requirements.txt
graphiti-core[google-genai]>=0.28.0
graphiti-mcp>=0.1.5
neo4j>=5.0.0
openpyxl>=3.1.0
pyyaml>=6.0
python-dotenv>=1.0.0
